saving segmentation to file feature done

+++ script_segemt_tool.py
@@ -38,7 +38,9 @@ class segment_obj:
         x_str = ' '.join([str(v) for v in self._xpixel])
         y_str = ' '.join([str(v) for v in self._ypixel])
 
-        str_segm = self._label + ':'
+        str_segm = self._label + ':' + x_str + ':' + y_str
+
+        return(str_segm)
 
 class bbox_object:
     def __init__(self,x,y,w,h,label):
@@ -119,12 +121,14 @@ class segment_gui:
         
         self.canvas.bind('<B1-Motion>',self.left_click_bb)
         self._kernel_value = 2
-        self.pixel_idx_to_plot = [[],[]] 
-        self.one_frame_pixels_idx = [[],[]]
-        self.bbox_list = []
+        self.pixel_bag = [[],[]] 
+        self.temp_pixel_bag = [[],[]]
+        self.elem_list = []
         self.selectedlabel = []
-        self.add_to_temp_bbox_bag([])
+        self.add_to_temp_obj_bag([])
         self.label_list = []
+        self.mode = self.set_mode('bbox')
+        self.unqlabel_list = []
 
         root = os.getcwd()
         self.segment_path = os.path.join(root,destin_path,'segment_labels')
@@ -144,6 +148,22 @@ class segment_gui:
         else:
             print ("Successfully created the directory %s " % self.bb_path)
 
+    def get_mode(self):
+        return(self.mode)
+
+    def set_mode(self,mode):
+
+        if mode  in ['bbox','segment']:
+            self.mode = mode
+            return(self.mode)
+        else: 
+            print("[WARNING] Mode does not exist!")
+            return(None)
+        
+    def plot_empty_canvas(self):
+        self.plot_canvas([],[])
+        self._window["-FILE LIST-"].update([])
+        self._window["-ELEM LIST-"].update([])
 
     def pointer_pixels(self,x,y,point_size):
 
@@ -174,15 +194,15 @@ class segment_gui:
         x_org,y_org,w,h = self._bb_tool.comp_bbox(self.pi,pf)
         bbox = bbox_object(x_org,y_org,w,h,[])
         # Add to temporary list
-        self.add_to_temp_bbox_bag(bbox)
+        self.add_to_temp_obj_bag(bbox)
         # Compute bounding box pixels 
         x_pixels,y_pixels = imgh.generate_bbox_pixels(x_org,y_org,w,h)
         # Clear the previous boudning box pixels from the temp bag
         self.clear_temp_pixel_bag()
         # Add the new pixels to temp bag
-        self.add_pixels_to_temp_bag(x_pixels,y_pixels)
+        self.add_to_temp_pixel_bag(x_pixels,y_pixels)
         # Get all bounding box pixels from the current frame
-        xp,yp = self.get_frame_pixels()
+        xp,yp = self.get_pixels()
         # Merge all pixels for plotting
         ypixels = np.concatenate((y_pixels,yp))
         xpixels = np.concatenate((x_pixels,xp))
@@ -200,37 +220,24 @@ class segment_gui:
 
         x_pixels,y_pixels = self.pointer_pixels(x,y,kernel_size)
 
-        self.add_pixels_to_temp_bag(x_pixels,y_pixels)
+        self.add_to_temp_pixel_bag(x_pixels,y_pixels)
+
+        segment = segment_obj([],x_pixels,y_pixels)
+
+        self.add_to_temp_obj_bag(segment)
         
         x_pixels,y_pixels = self.get_temp_pixels()
         self.plot_canvas(x_pixels,y_pixels)
-    
 
-    def reset_bbox_settings(self):
-        
-        self.bbox_list = []
-        self.label_list = []
-        self.reset_temp()
-
-    def get_pixels(self):
-        if np.array(self.pixel_idx_to_plot).size > 0:
-            xp,yp = np.array(self.pixel_idx_to_plot,dtype = int)
-            return(xp,yp)
-        else: 
-            return([],[])
-
-    def get_temp_pixels(self):
-        xp,yp = np.array(self.one_frame_pixels_idx,dtype = int)
-        return(xp,yp)
-    
-    def get_frame_pixels(self):
-        xp,yp = np.array(self.pixel_idx_to_plot,dtype = int)
-        return(xp,yp)
-
+  
     def plot_canvas(self,xp,yp):
 
         img = self._image_handler.get_curr_img()
         
+        if (img.size == 0): # <- change here
+            print("WARNING: No image found")
+            return(None)
+
         img[yp,xp, 0] = 255
         img[yp,xp, 1] = 0
         img[yp,xp, 2] = 0
@@ -239,53 +246,72 @@ class segment_gui:
 
         self._window["-IMAGE-"].update(data=imgbytes)
 
+    def clear_temp_obj_bag(self):
+        self.temp_obj_list = []
+
     def clear_temp_pixel_bag(self):
-        self.one_frame_pixels_idx = [[],[]]
+        self.temp_pixel_bag = [[],[]]
         
-    def reset_temp(self):
+    def clear_temp(self):
         self.clear_temp_pixel_bag()
-        self.clear_temp_bbox_bag()
+        self.clear_temp_obj_bag()
         self.first_bb_flag = False
+        #self.selectedlabel = []
 
     def clear_pixel_bag(self):
-        self.pixel_idx_to_plot = [[],[]]
+        self.pixel_bag = [[],[]]
+
+    def clear_obj_bag(self):
+        self.elem_list = []
+
+    def clear_elem(self):
+        self.clear_pixel_bag()
+        self.clear_obj_bag()
+        self.label_list = []
+        # self.clear_temp()
+
+    def clear_all(self):
+        self.clear_elem()
+        self.clear_temp()
+        self.unqlabel_list = []
 
     def add_to_pixel_bag(self,xidx,yidx):
 
-        self.pixel_idx_to_plot[0] = np.concatenate((self.pixel_idx_to_plot[0],xidx))
-        self.pixel_idx_to_plot[1] = np.concatenate((self.pixel_idx_to_plot[1],yidx))
+        self.pixel_bag[0] = np.concatenate((self.pixel_bag[0],xidx))
+        self.pixel_bag[1] = np.concatenate((self.pixel_bag[1],yidx))
 
-    def add_pixels_to_temp_bag(self,xidx,yidx):
+    def add_to_temp_pixel_bag(self,xidx,yidx):
 
-        self.one_frame_pixels_idx[0] = np.concatenate((self.one_frame_pixels_idx[0],xidx))
-        self.one_frame_pixels_idx[1] = np.concatenate((self.one_frame_pixels_idx[1],yidx))
+        self.temp_pixel_bag[0] = np.concatenate((self.temp_pixel_bag[0],xidx))
+        self.temp_pixel_bag[1] = np.concatenate((self.temp_pixel_bag[1],yidx))
 
-    def get_temp_bbox_bag(self):
-        return(self.select_bbox)
+    def add_to_temp_obj_bag(self,obj):
+        self.temp_obj_list = obj
 
-    def clear_temp_bbox_bag(self):
-        self.select_bbox = []
-
-    def add_to_temp_bbox_bag(self,bbox):
-        self.select_bbox = bbox
-
-    def add_to_bbox_list(self,bbox):
-        self.bbox_list.append(bbox)
+    def add_to_elem_list(self,bbox):
+        self.elem_list.append(bbox)
         self.label_list.append(bbox._label)
         return(self.label_list)
-    
-    def del_bbox(self):
-        self.clear_pixel_bag()
-        self.reset_bbox_settings()
-    
 
-    def get_bbox_list(self):
-        return(self.bbox_list)
+    def get_temp_obj_bag(self):
+        return(self.temp_obj_list)
+ 
+    def get_elem_list(self):
+        return(self.elem_list)
     
-    def save_bbox(self,bbox):
+    def get_pixels(self):
+        
+        xp,yp = np.array(self.pixel_bag,dtype = int)
+        return(xp,yp)
+        
+    def get_temp_pixels(self):
+        xp,yp = np.array(self.temp_pixel_bag,dtype = int)
+        return(xp,yp)
+
+    def save_temp_to_elem(self,obj):
 
         # Add the current bounding box to permanent list
-        label_list = self.add_to_bbox_list(bbox)
+        label_list = self.add_to_elem_list(obj)
         # Get the current bbox pixels 
         xp,yp = self.get_temp_pixels()
         # Update frame pixels with the new bbox pixels 
@@ -293,14 +319,19 @@ class segment_gui:
 
         return(label_list)
 
-    def save_bbox_to_file(self):
-        file_name = self._image_handler.get_curr_file_name()
-        file_path = os.path.join(self.bb_path,file_name+ '.txt')
+    def save_elem_to_file(self,path,filename):
+        
+        
+        if os.path.exists(path) == False:
+            print("[ERROR] File does not exist")
+            return()
+        
+        file_path = os.path.join(path,filename)
 
         f = open(file_path,'w')
-        bbox_list = self.get_bbox_list()
-        for bbox in bbox_list:
-            f.write(bbox.conv_to_str() + '\n')
+        elem_list = self.get_elem_list()
+        for elem in elem_list:
+            f.write(elem.conv_to_str() + '\n')
         f.close()
 
     def load_kernel_value(self,value):
@@ -352,10 +383,10 @@ class segment_gui:
             elif event == "-OK-":
 
                 label_name = values["-LABEL-"]
-                label_list = self._image_handler.appendLabel(label_name)
+                self.unqlabel_list.append(label_name)
                 
                 try:
-                    self._window["-FILE LIST-"].update(label_list)
+                    self._window["-FILE LIST-"].update(self.unqlabel_list)
                 except:
                     pass
 
@@ -371,8 +402,9 @@ class segment_gui:
             
             elif event == "-NEXT-":
                 # NEXT button was pressed: show next image
-                self.clear_pixel_bag()
-                self.reset_bbox_settings()
+                #self.clear_pixel_bag()
+                #self.clear_obj_bag()
+                self.clear_all()
                 self._window["-ELEM LIST-"].update([])
 
                 imgbytes = self._image_handler.get_next_img() # Get next image (as byte format) in line 
@@ -391,6 +423,9 @@ class segment_gui:
                 self._segment_tool_flag = False
                 
                 self.canvas.bind('<B1-Motion>',self.left_click_bb)
+                self.set_mode('bbox')
+                self.clear_all()
+                self.plot_empty_canvas()
 
             elif event == "-RADIO2-":
             # Selection of segmentation tool
@@ -399,33 +434,57 @@ class segment_gui:
                 self._segment_tool_flag = True
                 #self.canvas.unbind("<Button-1>")
                 self.canvas.bind('<B1-Motion>',self.left_click_segment)
+                self.set_mode('segment')
+
+                self.clear_all()
+                self.plot_empty_canvas()
             
-            elif  event == "-SAVE ELEMT-":
+            elif  event == "-SAVE OBJ-":
                 
-                temp_bag = self.get_temp_bbox_bag()
+                #print("[DEBUG] I'm SAVE obj")
+                temp_bag = self.get_temp_obj_bag()
                 label = self.selectedlabel
 
                 if label != [] and temp_bag != []:
-                    bbox = self.select_bbox
-                    bbox._label = label
-                    label_list = self.save_bbox(bbox)
-                    self.reset_temp()
+                    
+                    #if self.get_mode() == 'bbox':
+                        #print("[DEBUG] I'm SAVE obj")
+                    elem = temp_bag
+                    elem._label = label
+
+                    label_list = self.save_temp_to_elem(elem)
+                    #elif self.get_mode == 'segment':
+                        
+
+                    self.clear_temp()
                     self._window["-ELEM LIST-"].update(label_list)
 
-            elif  event == "-DEL ELEMT-":
+            elif  event == "-DEL OBJ-":
 
-                self.reset_temp()
+                self.clear_temp()
                 xp,yp = self.get_pixels()
                 self.plot_canvas(xp,yp)
             
             elif event == "-DEL-":
-                self.del_bbox()
+
+                self.clear_elem()
+                self.clear_temp()
+
                 self._window["-ELEM LIST-"].update([])
                 xp,yp = self.get_pixels()
                 self.plot_canvas(xp,yp)
 
             elif event == "-SAVE-":
-                self.save_bbox_to_file()
+                
+                file_name = self._image_handler.get_curr_file_name()
+                file_name += '.txt'
+                
+                if self.mode == 'bbox':
+                    path = self.bb_path
+                elif(self.mode == 'segment'): 
+                    path = self.segment_path
+
+                self.save_elem_to_file(path,file_name)
                 
         self._window.close()
 
@@ -451,8 +510,8 @@ file_list_column = [
         sg.Listbox(values=[], enable_events=True, size=(20, 10), key="-FILE LIST-")
     ],
     [
-        sg.Button('SAVE ELEMT',enable_events=True, key="-SAVE ELEMT-"),
-        sg.Button('DEL ELEMT',enable_events=True, key="-DEL ELEMT-")
+        sg.Button('SAVE OBJ',enable_events=True, key="-SAVE OBJ-"),
+        sg.Button('DEL OBJ',enable_events=True, key="-DEL OBJ-")
     ],
     [
         sg.Text("Elemtents")
